Added config watcher (#1)
Added Kubernetes configuration watcher. It uses the Kubernetes API to watch for
changes on ConfigMap or Secret resources and create files in the target
directory corresponding to the resource's `data` field. Multiple Kubernetes 
namespaces can be watched for changes. An optional webhook URL is invoked
when a change in resource results in a file content change in the target directory.
There are no guards for duplicate keys in ConfigMap resources.

Added `Dockerfile` and GitHub Actions to run the available tests.